libs2eplugins/Requiem: implement BinaryExprIterator library
This is Requiem's extension to klee.

Given an expression tree (where each node of the tree is either a BinaryExpr or ConstantExpr),
this forward iterator enables the user to easily traverse through the tree in the desired
traversal order (i.e., preorder, inorder, and postorder) which can be specified via
template argument `S`.

Example usage:

ref<Expr> root = ...;

for (auto it = BinaryExprIterator<IterStrategy::INORDER>::begin(root);
     it != decltype(it)::end();
     it++) {
     // Do something with *it, where *it is of type ref<Expr>.
}

Signed-off-by: Marco Wang <m.aesophor@gmail.com>